Eliminate need for module Sum.
Record the number of resources used by the original version abs16.v
which relies on a prefix adder for its implementation. These numbers
come from Vivado 2018.3 for the Digilent BASYS 3 board and may not match
the values reported by other versions of Vivado or for other FPGAs. The
information is provided to give a relative, not absolute, measure of how
efficient the orignal version of the abs16 module is compared with the
final version.