Replace references to P[i] with xorA[i]
P[i] is also known as Pi:i (P sub i:i).

Update the resources file to have a record of how the final version of
the abs16 module resource consumption compares with the original
version which relied on a full prefix adder.

Add a README.md file in preparation to upload the repository to GitHub.